+++ sharding-core/sharding-core-entry/src/main/java/org/apache/shardingsphere/core/shard/result/ShardingExecutionContext.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.core.shard.result;
 import com.google.common.base.Optional;
 import org.apache.shardingsphere.core.route.router.sharding.keygen.GeneratedKey;
 import org.apache.shardingsphere.sql.parser.relation.statement.SQLStatementContext;
-import org.apache.shardingsphere.underlying.route.context.ExecutionContext;
+import org.apache.shardingsphere.underlying.executor.context.ExecutionContext;
 
 /**
  * Sharding execution context.
