+++ sharding-core/sharding-core-execute/src/main/java/org/apache/shardingsphere/sharding/execute/sql/StatementExecuteUnit.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.sharding.execute.sql;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.underlying.executor.constant.ConnectionMode;
-import org.apache.shardingsphere.underlying.route.context.ExecutionUnit;
+import org.apache.shardingsphere.underlying.executor.context.ExecutionUnit;
 
 import java.sql.Statement;
