+++ sharding-core/sharding-core-execute/src/main/java/org/apache/shardingsphere/sharding/execute/sql/prepare/SQLExecutePrepareCallback.java
@@ -19,7 +19,7 @@ package org.apache.shardingsphere.sharding.execute.sql.prepare;
 
 import org.apache.shardingsphere.underlying.executor.constant.ConnectionMode;
 import org.apache.shardingsphere.sharding.execute.sql.StatementExecuteUnit;
-import org.apache.shardingsphere.underlying.route.context.ExecutionUnit;
+import org.apache.shardingsphere.underlying.executor.context.ExecutionUnit;
 
 import java.sql.Connection;
 import java.sql.SQLException;
