+++ sharding-core/sharding-core-execute/src/test/java/org/apache/shardingsphere/sharding/execute/sql/execute/SQLExecuteCallbackTest.java
@@ -21,8 +21,8 @@ import lombok.SneakyThrows;
 import org.apache.shardingsphere.underlying.executor.constant.ConnectionMode;
 import org.apache.shardingsphere.underlying.common.database.type.DatabaseTypes;
 import org.apache.shardingsphere.sharding.execute.sql.StatementExecuteUnit;
-import org.apache.shardingsphere.underlying.route.context.ExecutionUnit;
-import org.apache.shardingsphere.underlying.route.context.SQLUnit;
+import org.apache.shardingsphere.underlying.executor.context.ExecutionUnit;
+import org.apache.shardingsphere.underlying.executor.context.SQLUnit;
 import org.apache.shardingsphere.spi.database.metadata.DataSourceMetaData;
 import org.junit.Before;
 import org.junit.Test;
